+++ ui/service/src/Routes/projects/projectId/_snapshotId/index.tsx
@@ -1,0 +1,21 @@
+import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui-lib/routes-components/snapshotId/data'
+import { api } from 'api/RemoteApi'
+
+const { loader } = injectAPI({ api })
+
+export default {
+  path: ':snapshotId',
+  lazy: async () => {
+    const { SnapshotTemplate, ...rest } = await import(
+      'evidently-ui-lib/routes-components/snapshotId'
+    )
+
+    const Component = () => {
+      return <SnapshotTemplate api={api} />
+    }
+
+    return { Component, ...rest }
+  },
+  loader
+} satisfies RouteObject
