+++ ui/service/src/Routes/projects/projectId/test_suites/index.tsx
@@ -3,6 +3,5 @@ import { redirect } from 'react-router-dom'
 
 export default {
   path: 'test_suites/*',
-  loader: ({ request }) =>
-    redirect(request.url.replace('/test_suites', '/test-suites'), { status: 301 })
+  loader: ({ request }) => redirect(request.url.replace('/test_suites', '/test-suites'))
 } satisfies RouteObject
